breaking chance and breaking parts for asteroids
