+++ Assets/Mika/Prefabs/Asteroids/Brown/BreakableAsteroidBrown_1.prefab
@@ -13,7 +13,7 @@ GameObject:
   - component: {fileID: 7973108897030963214}
   - component: {fileID: -6164670779606190612}
   - component: {fileID: -5968271165429275820}
-  - component: {fileID: -9062237917773800952}
+  - component: {fileID: 4215582034946877954}
   m_Layer: 0
   m_Name: BreakableAsteroidBrown_1
   m_TagString: Untagged
@@ -163,7 +163,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 82c90de875616354f978095469a7bc4c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &-9062237917773800952
+--- !u!114 &4215582034946877954
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -172,13 +172,15 @@ MonoBehaviour:
   m_GameObject: {fileID: 8619275650118808469}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 06d66fc990780f340a9e72c56bebda1e, type: 3}
+  m_Script: {fileID: 11500000, guid: cc857adb05e1d9f4b80d59de01a29605, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   m_lives: 4
   m_speed: 0.4
   enemyData: {fileID: 11400000, guid: 18822097bec314245b53216cee6e6a74, type: 2}
-  smallAsteroidPrefab: {fileID: 1458167018991908628, guid: 373ec56cbf950b449a311675ca8a2532, type: 3}
+  smallAsteroidPrefabs: []
+  breakingParts: 3
+  breakingChance: 100
 --- !u!1001 &5103495211701118203
 PrefabInstance:
   m_ObjectHideFlags: 0
